requirements.txt
Flask==2.3.3
celery==5.3.4
redis==5.0.1
boto3==1.34.72
python-dotenv==1.0.1
flask_sqlalchemy
flask-cors
